Document and refactoring.
