+++ datashuttle/tui/tabs/transfer_status_tree.py
@@ -1,0 +1,145 @@
+import os
+from pathlib import Path
+
+from rich.style import Style
+from rich.text import Text
+from textual.widgets._directory_tree import DirEntry
+from textual.widgets._tree import TOGGLE_STYLE, TreeNode
+
+from datashuttle.configs import canonical_folders
+from datashuttle.tui.custom_widgets import (
+    CustomDirectoryTree,
+)
+from datashuttle.utils.rclone import get_local_and_central_file_differences
+
+
+class TransferStatusTree(CustomDirectoryTree):
+    """
+    A directorytree in which the nodes are styled depending on their
+    transfer status. e.g. indicates whether files are changed between
+    local or central, or appear in local only.
+    """
+
+    def __init__(self, parent_tab, project, id=None):
+        super(TransferStatusTree, self).__init__(
+            project.get_local_path(), id=id
+        )
+
+        self.parent_tab = parent_tab
+        self.project = project
+        self.update_transfer_diffs()
+
+    def on_mount(self):
+        self.transfer_paths = self.update_local_transfer_paths()
+
+    def update_local_transfer_paths(self):
+        """
+        Compiles a list of all project files and paths.
+        """
+        paths_list = []
+
+        for top_level_folder in canonical_folders.get_top_level_folders():
+            walk_paths = os.walk(
+                (self.project.get_local_path() / top_level_folder).as_posix()
+            )
+            for path in walk_paths:
+                paths_list.append(Path(path[0]))
+                if path[2]:
+                    paths_list.extend(
+                        [Path(f"{path[0]}/{file}") for file in path[2]]
+                    )
+
+        self.transfer_paths = paths_list
+
+    def update_transfer_diffs(self):
+        """
+        Updates the transfer diffs used to style the DirectoryTree.
+
+        Use `init` when the widget is initialised, because
+        #transfer_toplevel_radiobutton is not yet available and
+        by default we set to 'all'.
+        """
+        self.transfer_diffs = get_local_and_central_file_differences(
+            self.project.cfg, all_top_level_folder=True
+        )
+
+    def update_transfer_tree(self):
+        """
+        Updates tree styling to reflect the current TUI state
+        and project transfer status.
+        """
+        self.transfer_paths = self.update_local_transfer_paths()
+        self.update_transfer_diffs()
+        self.reload()
+
+    def format_transfer_label(self, node_label, node_path):
+        """
+        Takes nodes being formatted using `render_label` and applies custom
+        formatting according to the node's transfer status.
+        """
+
+        node_relative_path = node_path.as_posix().replace(
+            f"{self.project.cfg['local_path'] .as_posix()}/", ""
+        )
+
+        # Checks whether the current node's file path is staged for transfer
+        if node_path in self.transfer_paths:
+            # Sets sub- and ses-level folders to orange if files within have changed
+            # fmt: off
+            if node_relative_path in self.transfer_diffs["same"]:
+                pass
+            elif node_relative_path in self.transfer_diffs["different"] or any([node_relative_path in file for file in self.transfer_diffs["different"]]):
+                node_label.stylize_before("gold3")
+            elif node_relative_path in self.transfer_diffs["local_only"] or any([node_relative_path in file for file in self.transfer_diffs["local_only"]]):
+                node_label.stylize_before("green3")
+            elif node_label.plain in self.transfer_diffs["error"] or any([node_relative_path in file for file in self.transfer_diffs["error"]]):
+                node_label.stylize_before("bright_red")
+            # fmt: on
+
+    # Overridden Methods
+    # ----------------------------------------------------------------------------------
+
+    def render_label(
+        self, node: TreeNode[DirEntry], base_style: Style, style: Style
+    ) -> Text:
+        """
+        Extends the `DirectoryTree.render_label()` method to allow
+        custom styling of file nodes according to their transfer status.
+        """
+
+        node_label = node._label.copy()
+        node_label.stylize(style)
+
+        node_path = node.data.path
+
+        if node._allow_expand:
+            prefix = (
+                "📂 " if node.is_expanded else "📁 ",
+                base_style + TOGGLE_STYLE,
+            )
+            node_label.stylize_before(
+                self.get_component_rich_style(
+                    "directory-tree--folder", partial=True
+                )
+            )
+        else:
+            prefix = (
+                "📄 ",
+                base_style,
+            )
+            node_label.stylize_before(
+                self.get_component_rich_style(
+                    "directory-tree--file", partial=True
+                ),
+            )
+            node_label.highlight_regex(
+                r"\..+$",
+                self.get_component_rich_style(
+                    "directory-tree--extension", partial=True
+                ),
+            )
+
+        self.format_transfer_label(node_label, node_path)
+
+        text = Text.assemble(prefix, node_label)
+        return text
